finle upload

+++ routes/dishRouter.js
@@ -9,6 +9,7 @@ dishRouter.use = bodyParser.json();
 dishRouter.route('/')
 .get((req,res,next) => {
     Dishes.find({})
+    .populate('comments.author')
     .then((dishes) => {
         res.statusCode =200;
         res.setHeader('Content-Type', 'application/json');
@@ -17,6 +18,7 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Dishes.create(req.body)
     .then((dish) =>{
         console.log('Dish created', dish);
@@ -27,10 +29,12 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     res.statusCode=403;
     res.end('PUT opeation not supported on dishes');
 })
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -43,6 +47,7 @@ dishRouter.route('/')
 dishRouter.route('/:dishId')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) =>{
         console.log('Dish created', dish);
         res.statusCode = 200;
@@ -52,10 +57,12 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body,
     }, {new: true})
@@ -67,6 +74,7 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((dish) =>{
         res.statusCode = 200;
@@ -79,6 +87,7 @@ dishRouter.route('/:dishId')
 dishRouter.route('/:dishId/comments')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if (dish != null){
             res.statusCode =200;
@@ -97,12 +106,17 @@ dishRouter.route('/:dishId/comments')
     Dishes.findById(req.params.dishId)
     .then((dish) =>{
         if (dish != null){
+            req.body.author = req.user._id;
             dish.comments.push(req.body);
             dish.save()
             .then((dish) => {
-                res.statusCode =200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then ((dish) =>{
+                    res.statusCode =200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);
+                })
             }, (err) => next(err));
         } 
         else {
@@ -126,9 +140,13 @@ dishRouter.route('/:dishId/comments')
             }
             dish.save()
             .then((dish) => {
-                res.statusCode =200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then ((dish) =>{
+                    res.statusCode =200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);
+                })
             })
         } 
         else {
@@ -143,6 +161,7 @@ dishRouter.route('/:dishId/comments')
 dishRouter.route('/:dishId/comments/:commentId')
 .get((req,res,next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) =>{
         if (dish != null && dish.comments.id(req.params.commentId) != null){
             res.statusCode =200;
@@ -168,20 +187,36 @@ dishRouter.route('/:dishId/comments/:commentId')
 })
 .put(authenticate.verifyUser, (req,res,next) => {
     Dishes.findById(req.params.dishId)
-    .then((dish) =>{
+    .then((dish) =>{ 
+        
         if (dish != null && dish.comments.id(req.params.commentId) != null){
-            if (req.body.rating) {
-                dish.comments.id(req.params.commentId).rating = req.body.rating;
+            console.log("id of user: ", req.user._id);
+        console.log('id of comment: ', dish.comments.id(req.params.commentId).author);
+            if(dish.comments.id(req.params.commentId).author.toString() != req.user._id.toString()){
+                err = new Error('User: ' + req.user.username +  ' is not allowed to edit that comment');
+                err.status = 401;        
+                return next(err);
+                }
+            else{
+                console.log("match found");
+
+                if (req.body.rating) {
+                    dish.comments.id(req.params.commentId).rating = req.body.rating;
+                }
+                if (req.body.comment){
+                    dish.comments.id(req.params.commentId).comment = req.body.comment;
+                }
+
             }
-            if (req.body.comment){
-                dish.comments.id(req.params.commentId).comment = req.body.comment;
-            }
-            
             dish.save()
             .then((dish) => {
-                res.statusCode =200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish) => {
+                    res.statusCode =200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);
+                })
             }, (err) => next(err));
         } 
         else if (dish == null) {
