+++ routes/leaderRouter.js
@@ -17,6 +17,7 @@ leaderRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Leaders.create(req.body)
     .then((leader) =>{
         console.log('Leader created', leader);
@@ -27,10 +28,12 @@ leaderRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     res.statusCode=403;
     res.end('PUT opeation not supported on leaders/');
 })
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Leaders.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -52,10 +55,12 @@ leaderRouter.route('/:leaderId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     res.statusCode = 403;
     res.end('POST operation not supported on /leaders/'+ req.params.leaderId);
 })
 .put(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Leaders.findByIdAndUpdate(req.params.leaderId, {
         $set: req.body,
     }, {new: true})
@@ -67,6 +72,7 @@ leaderRouter.route('/:leaderId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req,res,next) => {
+    authenticate.verifyAdmin(req,next);  
     Leaders.findByIdAndRemove(req.params.leaderId)
     .then((leader) =>{
         res.statusCode = 200;
